Feature: delete TextHandler abstract class. Refactore: something
